Lots of meta programming
