+++ site/src/App/App.tsx
@@ -17,7 +17,7 @@ import foundations from './routes/foundations';
 import examples from './routes/examples';
 import components from './routes/components';
 import releaseNotes from './routes/releases';
-import { AppMeta } from './AppMeta';
+import { AppMeta } from './Seo/AppMeta';
 
 const CustomLink = makeLinkComponent(
   ({ href, rel, onClick, ...restProps }, ref) =>
