+++ site/src/App/ComponentDoc/ComponentDoc.tsx
@@ -20,6 +20,7 @@ import { useConfig } from '../ConfigContext';
 import { getHistory } from '../Updates';
 import { Markdown } from '../Markdown/Markdown';
 import { Navigation, NavigationItem } from './Navigation/Navigation';
+import { PageTitle } from '../Seo/PageTitle';
 
 const handler = () => {
   /* No-op for docs examples */
@@ -128,6 +129,7 @@ export const ComponentDoc = ({
       </Stack>
       <Switch>
         <Route exact path={`/components/${componentName}`}>
+          <PageTitle title={componentName} />
           <Stack space="xxlarge">
             {docs.description}
 
@@ -191,6 +193,8 @@ export const ComponentDoc = ({
           </Stack>
         </Route>
         <Route path={`/components/${componentName}/props`}>
+          <PageTitle title={`${componentName} Props`} />
+
           <Stack space="xxlarge">
             {Array.isArray(propsToDocument) ? (
               propsToDocument.map((c) => (
@@ -219,6 +223,8 @@ export const ComponentDoc = ({
           </Stack>
         </Route>
         <Route path={`/components/${componentName}/releases`}>
+          <PageTitle title={`${componentName} Releases`} />
+
           <Stack space="xlarge">
             {history.length > 0 ? (
               history.map((item, index) => (
