+++ site/src/App/Seo/favicon-32x32.png
