+++ site/src/App/Seo/favicon-96x96.png
