+++ site/src/App/routes/releases/index.tsx
@@ -1,15 +1,19 @@
-import React from 'react';
+import React, { Fragment } from 'react';
 
 // @ts-expect-error
 import changelogContent from '!!raw-loader!../../../../../CHANGELOG.md';
 
 import { Page } from '../../../types';
 import { Markdown } from '../../Markdown/Markdown';
+import { PageTitle } from '../../Seo/PageTitle';
 
 const Changelog = () => (
-  <Markdown>
-    {changelogContent.replace('# braid-design-system', '# Releases')}
-  </Markdown>
+  <Fragment>
+    <PageTitle title="Releases" />
+    <Markdown>
+      {changelogContent.replace('# braid-design-system', '# Releases')}
+    </Markdown>
+  </Fragment>
 );
 
 const page: Page = {
